src/requirements_sam.txt
opencv-python>=4.5.0
matplotlib>=3.3.0
numpy>=1.19.0
Pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
git+https://github.com/facebookresearch/segment-anything.git 